DemoServer/requirements.txt
asgiref==3.7.2
black==23.7.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.3
Django==4.1
django-cors-headers==4.2.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
docopt==0.6.2
et-xmlfile==1.1.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
mypy-extensions==1.0.0
numpy==1.25.2
openapi-codec==1.3.2
openpyxl==3.1.2
packaging==23.1
pandas==2.1.0
pathspec==0.11.2
pipreqs==0.4.13
platformdirs==3.10.0
pycparser==2.21
PyJWT==2.8.0
PyMySQL==1.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.4
uWSGI==2.0.22
yarg==0.1.9
